# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Banco D$ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/password:
    patch:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/account:
    delete:
      operationId: deleteAccount
      tags: [users]
      summary: Delete account (requires zero balance in all reserves and main balance)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Non-zero balance or wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        balance:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - username
        - balance
        - createdAt
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
      required:
        - id
        - username
    UpdateProfileRequest:
      type: object
      properties:
        avatarUrl:
          type: string
          nullable: true
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
        confirmNewPassword:
          type: string
      required:
        - currentPassword
        - newPassword
        - confirmNewPassword
    DeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password
...